pyproject.toml
[project]
name = "relay-it-api"
version = "0.1.0"
description = "Relay it! API - Screenshot analysis with Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "google-generativeai>=0.8.0",
]
